Addressed Oliver's comments. Factored out some common functionality.
